entities: moved types to entities directory
