+++ src/entities/journal/journal-entry.ts
